pfade angepasst, hasAuthorities im Backend angepasst, config Headers...

See merge request stellabrettschneider/cookingschool!46

+++ cookingSchoolBackend/src/main/java/Cooking/School/Project/cookingSchool/Services/UserService.java
@@ -7,11 +7,12 @@ import Cooking.School.Project.cookingSchool.repository.CourseRepository;
 import Cooking.School.Project.cookingSchool.repository.UserRepository;
 import Cooking.School.Project.cookingSchool.restapi.dto.UserResponse;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DuplicateKeyException;
+
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @Service
@@ -26,19 +27,49 @@ public class UserService {
     public UserService() {
     }
 
+/*
     public User addUser(User user) {
         userRepository.save(user);
         return user;
-    }
+    }*/
+    /* war neu
+    public User addUser(User user){
+        Optional<User> existingUserByEmail = userRepository.findUserByEmail(user.getEmail());
+        Optional<User> existingUserByUsername = userRepository.findUserByUsername(user.getUsername());
+
+        if (existingUserByEmail.isPresent()) {
+            String errorMessage = "Die E-Mail-Adresse '" + user.getEmail() + "' ist bereits registriert.";
+            throw new DuplicateKeyException(errorMessage);
+        } else if (existingUserByUsername.isPresent()) {
+            String errorMessage = "Der Benutzername '" + user.getUsername() + "' ist bereits registriert.";
+            throw new DuplicateKeyException(errorMessage);
+        } else {
+            return userRepository.save(user);
+        }
+    }*/
 
     public UserResponse getUserById(Long id) {
         User user = userRepository.findById(id)
                 .orElseThrow(() -> new UserNotFoundException("User not found " +id));
         return UserResponse.fromUser(user);
     }
-
+    /*
+    public BasicUserResponse getUserById(Long id) {
+        User user = userRepository.findById(id)
+                .orElseThrow(() -> new UserNotFoundException("User not found " + id));
+        return UserMapper.INSTANCE.toUserResponse(user);
+    }*/
     public void deleteUserById(Long id) {
-        userRepository.deleteById(id);
+        if (id == null) {
+            throw new PrimaryIdNullOrEmptyException("User Id is null or empty " + id);
+        }
+        Optional<User> optionalUser = userRepository.findById(id);
+
+        if (optionalUser.isPresent()) {
+            userRepository.deleteById(id);
+        } else {
+            throw new UserNotFoundException("User not found with ID: " + id);
+        }
     }
 
 
@@ -53,6 +84,7 @@ public class UserService {
         existingUser.setFirstname(updatedUser.getFirstname());
         existingUser.setLastname(updatedUser.getLastname());
         existingUser.setAddress(updatedUser.getAddress());
+        existingUser.setMobile(updatedUser.getMobile());
         existingUser.setPassword(updatedUser.getPassword());
         existingUser.setUsername(updatedUser.getUsername());
         existingUser.setAdmin(updatedUser.isAdmin());
@@ -61,33 +93,31 @@ public class UserService {
 
         return existingUser;
 
-
-    }
-    @Transactional
-    public User editUser(Long userId, User updatedUser) throws PrimaryIdNullOrEmptyException {
-        if (userId == null) {
-            throw new PrimaryIdNullOrEmptyException("User Id is null or empty");
-        }
-        User existingUser = userRepository.findById(userId)
-                .orElseThrow(() -> new UserNotFoundException("User not found"));
-
-        existingUser.setFirstname(updatedUser.getFirstname());
-        existingUser.setLastname(updatedUser.getLastname());
-        existingUser.setAddress(updatedUser.getAddress());
-        existingUser.setPassword(updatedUser.getPassword());
-        existingUser.setUsername(updatedUser.getUsername());
-        existingUser.setAdmin(updatedUser.isAdmin());
-
-        userRepository.save(existingUser);
-
-        return existingUser;
     }
 
+
+    /**
+     * Get Method to get a List of all Users
+     *
+     * @return
+     */
 
     public List<User> getAllUsers() {
-        return userRepository.findAll();
+        List<User> users = userRepository.findAll();
+        if (users.isEmpty()) {
+            throw new NoSuchElementException("No users found");
+        }
+        return users;
     }
 
+    /**
+     * book a course for a user and handle number of participants and save it in the database
+     * @param userId
+     * @param courseId
+     * @throws CourseNotFoundException
+     * @throws UserNotFoundException
+     * @throws MaxAttendantsReachedException when course is fully booked
+     */
     @Transactional
     public void bookCourse(Long userId, Long courseId) {
 
@@ -112,15 +142,27 @@ public class UserService {
         }
     }
 
-    public User registration(User user) throws DuplicateKeyException {
-        Optional<User> existingUser = userRepository.findUserByEmail(user.getEmail());
 
-        if (existingUser.isPresent()) {
-            throw new DuplicateKeyException(user.getEmail());
+    /**
+     * register a new user
+     * @param user
+     * @return DuplicateKeyException If the email address or username is already registered.
+     */
+    public User registration(User user){
+        Optional<User> existingUserByEmail = userRepository.findUserByEmail(user.getEmail());
+       Optional<User> existingUserByUsername = userRepository.findUserByUsername(user.getUsername());
+
+        if (existingUserByEmail.isPresent()) {
+            String errorMessage = "Die E-Mail-Adresse '" + user.getEmail() + "' ist bereits registriert.";
+            throw new DuplicateKeyException(errorMessage);
+        } else if (existingUserByUsername.isPresent()) {
+            String errorMessage = "Der Benutzername '" + user.getUsername() + "' ist bereits registriert.";
+            throw new DuplicateKeyException(errorMessage);
         } else {
             return userRepository.save(user);
         }
     }
+
 }
